src/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.8.2
python-dotenv==1.0.1
semantic-kernel==1.35.2
azure-search-documents==11.5.3
azure-core>=1.30.0
azure-cosmos>=4.7.0
pyodbc>=5.1.0

# Additional dependencies for the full backend
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
openai>=1.0.0
python-jose[cryptography]==3.3.0
python-multipart>=0.0.7
PyJWT>=2.8.0
requests>=2.31.0
httpx>=0.25.2
python-dateutil>=2.8.2
pytest==7.4.3
pytest-asyncio==0.21.1
cachetools>=5.3.0

# Azure AI Foundry integration
azure-ai-projects>=1.0.0b1